requirements.txt
natsort==8.4.0
numpy==1.26.4
opensimplex==0.4.5
Pillow==10.3.0
ursina==7.0.0
